Renamed: Base64Text to Base64Decoded, TextBase64 to Base64Encoded (#1004)
